System Damage w grze
-nie oneshotujemy wrogów
-wrogowie mają healthbar (kinda)
