+++ Assets/PreFab/pfProjectile.prefab
@@ -97,8 +97,9 @@ GameObject:
   - component: {fileID: 900874904887636218}
   - component: {fileID: 3927427240530742302}
   - component: {fileID: -2567369407421787585}
+  - component: {fileID: 1303106960607143170}
   m_Layer: 9
-  m_Name: pfFireball
+  m_Name: pfProjectile
   m_TagString: Weapon
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -139,7 +140,7 @@ Rigidbody2D:
   m_Material: {fileID: 0}
   m_IncludeLayers:
     serializedVersion: 2
-    m_Bits: 64
+    m_Bits: 0
   m_ExcludeLayers:
     serializedVersion: 2
     m_Bits: 776
@@ -160,9 +161,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 0}
-  body: {fileID: 0}
-  stats: {fileID: 0}
-  mousePos: {x: 0, y: 0, z: 0}
+  body: {fileID: -5772702193275167778}
 --- !u!61 &3927427240530742302
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -192,7 +191,7 @@ BoxCollider2D:
   m_CallbackLayers:
     serializedVersion: 2
     m_Bits: 4294967295
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0.0016737804, y: 0.002606988}
@@ -321,3 +320,16 @@ TrailRenderer:
   m_Autodestruct: 0
   m_Emitting: 1
   m_ApplyActiveColorSpace: 1
+--- !u!114 &1303106960607143170
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3621023564299458670}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac2da97659ffe0f4a9a85e0ac2b21dd6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  attackDamage: 1
