+++ Assets/PreFab/pfSlash.prefab
@@ -13,6 +13,7 @@ GameObject:
   - component: {fileID: 5343511410743520800}
   - component: {fileID: 4141074371457540291}
   - component: {fileID: 8438659673209803374}
+  - component: {fileID: -3100631860052788725}
   m_Layer: 9
   m_Name: pfSlash
   m_TagString: Weapon
@@ -137,7 +138,7 @@ BoxCollider2D:
   m_CallbackLayers:
     serializedVersion: 2
     m_Bits: 4294967295
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: 0, y: 0}
@@ -165,4 +166,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e8ba8cd5d83d5e6458aaece7bfce4e50, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  healthBar: {fileID: 0}
+--- !u!114 &-3100631860052788725
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4466358947862510512}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ac2da97659ffe0f4a9a85e0ac2b21dd6, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  attackDamage: 2
