+++ Assets/Scripts/Coin.cs
@@ -6,13 +6,6 @@ public class Coin : MonoBehaviour
 {
     private float despawnTime = 100f;
 
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
     // Update is called once per frame
     void Update()
     {
