+++ Assets/Scripts/Damage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac2da97659ffe0f4a9a85e0ac2b21dd6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
