+++ Assets/Scripts/Flipper2.cs
@@ -2,7 +2,6 @@ using UnityEngine;
 
 public class Flipper2 : MonoBehaviour
 {
-    [SerializeField]
     void Flip()
     {
         if (Input.GetAxis("Horizontal") < 0)
@@ -13,11 +12,6 @@ public class Flipper2 : MonoBehaviour
         {
             transform.rotation = Quaternion.Euler(0, 0, 0);
         }
-    }
-    // Start is called before the first frame update
-    void Start()
-    {
-
     }
 
     // Update is called once per frame
