+++ Assets/Scripts/HealthBar.cs
@@ -4,19 +4,19 @@ using UnityEngine.UI;
 public class HealthBar : MonoBehaviour
 {
 
-    public Slider slider;
-    public Gradient gradient;
+	public Slider slider;
+	public Gradient gradient;
 
 
-    public void SetMaxHealth(int health)
-    {
-        slider.maxValue = health;
-        slider.value = health;
-    }
+	public void SetMaxHealth(int health)
+	{
+		slider.maxValue = health;
+		slider.value = health;
+	}
 
-    public void SetHealth(int health)
-    {
-        slider.value = health;
-    }
+	public void SetHealth(int health)
+	{
+		slider.value = health;
+	}
 
 }
