+++ Assets/Scripts/Slash.cs
@@ -2,27 +2,20 @@ using UnityEngine;
 
 public class Slash : MonoBehaviour
 {
-    [SerializeField]
-    public HealthBar healthBar;
+    //public HealthBar healthBar;
     private float time = 1f;
-    static public int heal = 1;
 
-    private void OnCollisionEnter2D(Collision2D collision)
+    private void OnTriggerEnter2D(Collider2D collision)
     {
-        Vector3 deadPreFabCoords = collision.gameObject.transform.position;
-        Destroy(collision.gameObject);
         /* narazie zostawiam to wylaczone
         Stats.currentHp += heal;
         healthBar.SetHealth(Stats.currentHp);
         */
-        EnemySpawn.spawned--;
-        ExpManager.Instance.AddExp(Enemy.enemyExp);
-        FindAnyObjectByType<CoinDrop>().Drop(deadPreFabCoords);
     }
 
     void Start()
     {
-        healthBar = FindObjectOfType<HealthBar>();
+        //healthBar = FindObjectOfType<HealthBar>();
     }
 
     void Update()
